putting decimals in person and personSessionWeight (height and weight). DB: 0.75 Tests remain.

+++ src/server.cs
@@ -555,7 +555,7 @@ public class Server
 		return person;
 	}
 
-	private static void serverUploadPersonSessionIfNeeded(ChronojumpServer myServer, int personServerID, int sessionServerID, int weight)
+	private static void serverUploadPersonSessionIfNeeded(ChronojumpServer myServer, int personServerID, int sessionServerID, double weight)
 	{
 		myServer.UploadPersonSessionIfNeeded(personServerID, sessionServerID, weight);
 	}
